Edit module and class and Add filter method.
