+++ Gemfile
@@ -1,6 +1,5 @@
-# frozen_string_literal: true
 gem 'rubocop', '>= 1.0', '< 2.0'
 
-source "https://rubygems.org"
+source 'https://rubygems.org'
 
 # gem "rails"
