--- MyList.rb
@@ -1,12 +1,0 @@
-class MyList
-    include MyEnumerable
-  
-    def initialize(list)
-      @list = list
-    end
-  
-    def each
-      @list.each { |element| yield element }
-    end
-  end
-
